Add ability to clear cached firmware versions from DIS
Implements functions to clear only the firmware version field from
cached DIS info, forcing it to be re-read on next connection.

This solves the problem where:
1. Dongle caches MouthPad firmware version for fast reconnection
2. User updates MouthPad firmware directly via system BLE
3. Dongle incorrectly reports old cached version after reconnection

New functions:
- ble_dis_clear_cached_firmware_for_addr() - Clear for specific device
- ble_dis_clear_all_cached_firmware() - Clear for all bonded devices

New shell command:
- clearfwcache - Clears cached firmware for all bonds

This provides the foundation for a future protobuf message that
the app can send when initiating a firmware update flow.